# src/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Payment API (India)",
    "description": "Payment processing API for India – UPI, cards, net banking, wallets, cash on delivery. TrendRush-style schema (raw MySQL).",
    "version": "1.0.0",
    "contact": { "name": "API Support" }
  },
  "servers": [
    { "url": "http://localhost:3000", "description": "Local" }
  ],
...
  "paths": {
...
    "/api/orders/create": {
      "post": {
        "tags": ["Orders"],
        "summary": "Create order",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["userId", "items", "totalAmount", "shippingAddress", "billingAddress"],
                "properties": {
                  "userId": { "type": "string" },
                  "items": {
                    "type": "array",
                    "items": { "type": "object", "properties": { "name": {}, "quantity": {}, "price": {} } }
                  },
                  "totalAmount": { "type": "number" },
                  "shippingAddress": { "type": "object" },
                  "billingAddress": { "type": "object" },
                  "paymentMethod": { "type": "string" },
                  "gstNumber": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Order created; returns orderId (order_number), order_id, shippingAddressId, billingAddressId"
          },
          "400": { "description": "Validation failed" }
        }
      }
    },
    "/api/orders/{orderId}": {
      "get": {
        "tags": ["Orders"],
        "summary": "Get order by ID or order number",
        "parameters": [
          { "name": "orderId", "in": "path", "required": true, "schema": { "type": "string" }, "description": "Order number (ORD_...) or numeric order_id" }
        ],
        "responses": {
          "200": { "description": "Order and payment (if any)" },
          "404": { "description": "Order not found" }
        }
      }
    },
    "/api/orders/user/{userId}": {
      "get": {
        "tags": ["Orders"],
        "summary": "List orders by user",
        "parameters": [
          { "name": "userId", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "page", "in": "query", "schema": { "type": "integer", "default": 1 } },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 10 } }
        ],
        "responses": { "200": { "description": "Orders and pagination" } }
      }
    },
    "/api/orders/{orderId}/status": {
      "patch": {
        "tags": ["Orders"],
        "summary": "Update order status",
        "parameters": [
          { "name": "orderId", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["status"],
                "properties": { "status": { "type": "string" } }
              }
            }
          }
        },
        "responses": { "200": { "description": "Status updated" } }
      }
    },
    "/api/orders/{orderId}/cancel": {
      "patch": {
        "tags": ["Orders"],
        "summary": "Cancel order",
        "parameters": [
          { "name": "orderId", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": { "reason": { "type": "string" } }
              }
            }
          }
        },
        "responses": { "200": { "description": "Order cancelled" } }
      }
    },
...
  }
}